// front_src/src/model/user.ts
export class User {
    firstname:string;
    lastname:string;
    username: string;
    password: string;
    type: number;
    phone: number;
    email: string;
    approved: number;
    agname:string;
    State:string;
    City: string;
    Street :string;
    snumber :number;
    PIB: number;
    description:string;
    url:string;
    brojobjekata:number;
}